+++ StatelessCMS.php
@@ -3,7 +3,7 @@
  * StatelessCMS.php
  * 
  * @package StatelessCMS
- * @version 0.2.1
+ * @version 0.3.0
  */
 
 /**