+++ script/clean.sh
@@ -1,0 +1,2 @@
+rm -rf build/*.zip;
+rm -rf build/*.tar.gz;