+++ script/rebuild.sh
@@ -1,0 +1,2 @@
+./script/clean.sh;
+./script/build.sh;